+++ src/openaivec/embeddings.py
